Fix t_encrypt_compat ignoring frame size
t_encrypt_compat was missing a call to set frame size, so it was only
testing the default frame size. This also turned up a bug in the test
case itself resulting in livelock for some framesizes due to an
incorrect condition controlling when the full message length was set on
the session.