RFC: Caching CMM API
This is a first draft of an API for the caching CMM in C-land. This is more or
less a direct translation of the Java API, but with some small caveats;
notably, the extensible UsageHints structure is replaced by a simple timestamp,
and UsageStats is not virtualized. This trades off some future flexibility for
a substantial reduction in implementation complexity.