Local cache: Encrypt side
This change implements the encrypt side of the local cache. It also includes a
modified version of the cache headers previously seen in PR #65. Some basic
functional tests are in place, but more will be added in followup PRs or commits.

Note that the API has changed from PR #65; in particular, we no longer have a
vtable for entries, and instead put the entry methods on the cache object. This
is because putting the methods on the entry required a reference from the entry
to the parent cache, and the interaction between this reference, destruction of
the parent cache, and the cache mutex proved extremely complicated to manage.

Placing the methods on the cache object resolves these issues by requiring that
the cache remain alive until calling code is done using the entries.