Add codebuild logic for GCC 4.8 x86/x64 builds
This adds dockerfiles and codebuild buildspecs for a GCC 4.8 x86/x64 build on
Ubuntu Trusty. These are meant to serve as a template for future CI
configurations; in a future change, the existing codebuild logic under
build-tools/bin and buildspec.yml in the root will be merged with this logic
and moved into the codebuild directory as well.