Add new codebuild CI environments (#124)
* Reduce decryption vectors test time under REDUCE_TEST_ITERATIONS

* Add codebuild logic for GCC 4.8 x86/x64 builds

This adds dockerfiles and codebuild buildspecs for a GCC 4.8 x86/x64 build on
Ubuntu Trusty. These are meant to serve as a template for future CI
configurations; in a future change, the existing codebuild logic under
build-tools/bin and buildspec.yml in the root will be merged with this logic
and moved into the codebuild directory as well.

* Enable logging in the C++ integration test

* Emit valgrind suppressions on the console

* C++ integration test logging improvements to debug CI failures

* Adjust suppressions

* Enable valgrind suppresions on the preexisting Linux CI build

* dtor -> destructor

* Add some _s to variable names

* Bump up request timeouts

* Run the valgrind suite last

This is to see if the intermittent failures on codebuild are caused by valgrind
or inherent to the tests.