Refactor the materials cache API
Returning the materials and encryption context as part of the initial call to
the cache was awkward, as the caching CMM needs to decide whether it will
accept the entry based on the usage stats, and if it does not accept the entry,
it must perform a cache miss using the _original_ request encryption context.

To avoid needing to make extra copies of these contexts, we split the operation
into three steps: Searching for the entry (which is a common function for
encrypt and decrypt entries), checking/updating usage stats, and retrieving the
materials.