Simplify KMS keyring inheritance (#145)
* Simplify KMS Keyring inheritance

KmsKeyring class inherits directly from aws_cryptosdk_keyring.
Middle man class is eliminated. Also, providing a (struct aws_allocator *)
to the KmsKeyring class is now completely eliminated, because its memory
is entirely allocated by the allocator of the AWS C++ SDK using Aws::New
and Aws::Delete.

Previously the allocator pointer being stored was only being used in one
place, and it was being used incorrectly, as it was something that should
have been used by the request allocator. This has been corrected to use
the request allocator. This is the only functional change
in this commit.

This simplifies the construction of the KmsKeyring objects and it means
that all (struct aws_allocator *) appearances in the C++ code are only
for request allocators.

* Do all setup in constructor and get rid of unneeded Init functions