Remove KMS Client Cache object and refactor Client Suppliers
There is now a CachingClientSupplier which includes the cache as
part of that class, rather than requiring the user to configure
both supplier and cache separately.

Also changed default region handling. Now default region does not
need to be set when only key IDs in ARN format are provided.

Changed names of Set* and Append* functions in Builder to With*.

Removed a few tests which were no longer relevant.