Add some doc comments and README files describing the state of the CPP stuff
Note that the doxygen configuration appears to be a bit messed up, and promotes
the namespace docs to the index page of the generated doxygen; this is
something we should fix as part of our larger doxygen audit. For this PR,
please focus on whether you agree with the phrasing here.