Remove obsolete CI configuration
The CI configuration that used this was an early ubuntu LTS target we set up
before building our own docker images. Now that we're building a custom LTS
image using the logic under codebuild/, and now that this target has broken due
to changes on the ubuntu image side, this target has been disabled, and its
support files can be removed.