Add AWS C-Common submodule for CBMC proofs
This commit adds AWS C-Common [1], a dependency of E-SDK that is
required to build CBMC proofs, as a submodule to the repository.
Previously, the proof build process would clone the C-Common repository
if it was not already available, giving us less control over the
particular version of C-Common that was being used.

[1] https://github.com/awslabs/aws-c-common/