chore: run examples along with tests (#579)
This PR fixes some of the examples, and includes each example in the test suite. This fixes #552.

Operational note: In order to run the tests that make KMS calls, it is necessary to set some environment variable to KMS key ARNs (e.g. AWS_ENC_SDK_EXAMPLE_KEY_ARN). This PR configures CMake to print an error to the user if any required environment variables are unset. I have set these environment variables appropriately for our CI environments, so that the CI runs for this PR should also run the examples.